+++ doc/production-hosting.md
@@ -35,20 +35,20 @@ There are three public badge servers on OVH VPS’s.
 | [s1.shields-server.com][s1] | vps244529.ovh.net    | VPS  | 51.254.114.150 | Gravelines, France |
 | [s2.shields-server.com][s2] | vps117870.vps.ovh.ca | VPS  | 149.56.96.133  | Quebec, Canada     |
 
-- These are single-core virtual hosts with 2 GB RAM [VPS SSD 1]().
-- The Node version (v9.4.0 at time of writing) and dependency versions on the
+* These are single-core virtual hosts with 2 GB RAM [VPS SSD 1]().
+* The Node version (v9.4.0 at time of writing) and dependency versions on the
   servers can be inspected in Sentry, but only when an error occurs.
-- The servers use self-signed SSL certificates. ([#1460][issue 1460])
-- After accepting the certificate, you can debug an individual server using
+* The servers use self-signed SSL certificates. ([#1460][issue 1460])
+* After accepting the certificate, you can debug an individual server using
   the links above.
-- The scripts that start the server live in the [ServerScript][] repo. However
+* The scripts that start the server live in the [ServerScript][] repo. However
   updates must be pulled manually. They are not updated as part of the deploy process.
-- The server runs SSH.
-- Deploys are made using a git post-receive hook.
-- The server uses systemd to automatically restart the server when it crashes.
-- Provisioning additional servers is a manual process which is yet to been
+* The server runs SSH.
+* Deploys are made using a git post-receive hook.
+* The server uses systemd to automatically restart the server when it crashes.
+* Provisioning additional servers is a manual process which is yet to been
   documented.
-- The public servers _do not_ use docker. The `Dockerfile` is included for
+* The public servers _do not_ use docker. The `Dockerfile` is included for
   self-hosting (including on a Docker-capable PaaS).
 
 [s0]: https://s0.shields-server.com/index.html
@@ -67,12 +67,12 @@ Shields has mercifully little persistent state:
 2.  The analytics data is also saved on each server in JSON files on disk.
 3.  The server keeps a few caches in memory. These are neither persisted nor
     inspectable.
-    - The [request cache][]
-    - The [regular-update cache][]
-    - The [raster cache][]
+    * The [request cache][]
+    * The [regular-update cache][]
+    * The [raster cache][]
 
 [github auth admin endpoint]: https://github.com/badges/shields/blob/master/services/github/auth/admin.js
-[request cache]: https://github.com/badges/shields/blob/master/lib/request-handler.js#L29-L30
+[request cache]: https://github.com/badges/shields/blob/master/core/base-service/legacy-request-handler.js#L29-L30
 [regular-update cache]: https://github.com/badges/shields/blob/master/lib/regular-update.js
 [raster cache]: https://github.com/badges/shields/blob/master/gh-badges/lib/svg-to-img.js#L9-L10
 [oauth transfer]: https://developer.github.com/apps/managing-oauth-apps/transferring-ownership-of-an-oauth-app/
@@ -90,11 +90,11 @@ NODE_CONFIG_ENV=shields-io-production
 With that variable set, the server ([using `config`][config]) reads these
 files:
 
-- [`local-shields-io-production.yml`][local-shields-io-production.yml].
+* [`local-shields-io-production.yml`][local-shields-io-production.yml].
   This file contains secrets which are checked in with a deploy commit.
-- [`shields-io-production.yml`][shields-io-production.yml]. This file
+* [`shields-io-production.yml`][shields-io-production.yml]. This file
   contains non-secrets which are checked in to the main repo.
-- [`default.yml`][default.yml]`. This file contains defaults.
+* [`default.yml`][default.yml]`. This file contains defaults.
 
 [start-shields.sh]: https://github.com/badges/ServerScript/blob/master/start-shields.sh#L7
 [config]: https://github.com/lorenwest/node-config/wiki/Configuration-Files
@@ -109,9 +109,9 @@ The project ships with `dotenv`, however there is no `.env` in production.
 Sitting in front of the three servers is a Cloudflare Free account which
 provides several services:
 
-- Global CDN, caching, and SSL gateway for `img.shields.io`
-- Analytics through the Cloudflare dashboard
-- DNS hosting for `shields.io`
+* Global CDN, caching, and SSL gateway for `img.shields.io`
+* Analytics through the Cloudflare dashboard
+* DNS hosting for `shields.io`
 
 Cloudflare is configured to respect the servers' cache headers.
 
@@ -156,10 +156,10 @@ To deploy:
 3.  Run `npm ci`. **This is super important for the frontend build!**
 4.  Run `make deploy-s0` to make a canary deploy.
 5.  Check the canary deploy:
-    - [Visit the server][s0]. Don't forget that most of the preview badges
+    * [Visit the server][s0]. Don't forget that most of the preview badges
       are static!
-    - Look for errors in [Sentry][].
-    - Keep an eye on the [status page][status].
+    * Look for errors in [Sentry][].
+    * Keep an eye on the [status page][status].
 6.  After a little while (usually 10–60 minutes), finish the deploy:
     `make push-s1 push-s2 deploy-gh-pages`.
 
@@ -213,9 +213,9 @@ the server. It's generously donated by [Sentry][sentry home]. We bundle
 
 Request performance is monitored in two places:
 
-- [Status][] (using [UptimeRobot][])
-- [Server metrics][] using Prometheus and Grafana
-- [@RedSparr0w's monitor][monitor] which posts [notifications][] to a private
+* [Status][] (using [UptimeRobot][])
+* [Server metrics][] using Prometheus and Grafana
+* [@RedSparr0w's monitor][monitor] which posts [notifications][] to a private
   [#monitor chat room][monitor discord]
 
 Overall server performance is monitored using Prometheus and Grafana.
